Store settings in the DB instead of a config file
Use a simple key value table for config instead of a config file.  This
allows us to easily test the app with the :memory: sqlite table, lets
us remove a storage path we don't need anymore, allows backing up
everything with a single file, and deletes lots of code.

Use the name 'Settings' instead of 'Config' so it's clear that we aren't
talking about application config.