Use Closure event callbacks
Apparently Laravel will resolve the subscriber from the container each
time it calls an event listener.  Because a fresh instance is resolved
the environment property is null and the callback never runs.  To
prevent this we use a closure which forces the dispatcher to use our
existing instance instead of resolving the class 3 separate times.