Adds #12 default project.
First we check the argument provided to the command.
If none is given we check the env setting using `getenv`. Env is read from the `.hours` file
The last check performed is the settings database.

If no project is given we throw an error to the screen informing the user they must provide a project